# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Royal Ludo Arena API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  referralCode: string;
  wins: number;
  losses: number;
  banned: boolean;
  isAdmin: boolean;
  createdAt: string;
}

export interface Wallet {
  userId: number;
  mainBalance: number;
  winningBalance: number;
  bonusBalance: number;
  totalBalance: number;
}

export interface Room {
  id: number;
  name: string;
  label: string;
  entryFee: number;
  winnerPrize: number;
  playerCount: number;
  tagline: string;
  color: string;
  active: boolean;
}

export interface Match {
  id: string;
  roomId: number;
  roomName: string;
  player1Name: string;
  /** @nullable */
  player2Name?: string | null;
  /** @nullable */
  winnerName?: string | null;
  entryFee: number;
  prize: number;
  status: string;
  createdAt: string;
}

export interface MatchStatus {
  /** @nullable */
  matchId?: string | null;
  status: string;
  /** @nullable */
  opponent?: string | null;
  roomId: number;
  message: string;
}

export interface Deposit {
  id: number;
  userId: number;
  userName: string;
  amount: number;
  trxId: string;
  method: string;
  phone: string;
  status: string;
  createdAt: string;
}

export interface Withdrawal {
  id: number;
  userId: number;
  userName: string;
  amount: number;
  method: string;
  phone: string;
  status: string;
  createdAt: string;
}

export interface Notification {
  id: number;
  /** @nullable */
  userId?: number | null;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export type ReferralInfoReferredUsersItem = {
  name: string;
  joinedAt: string;
};

export interface ReferralInfo {
  referralCode: string;
  referralBonus: number;
  totalReferred: number;
  totalEarned: number;
  referredUsers: ReferralInfoReferredUsersItem[];
}

export interface AppSettings {
  telegramLink: string;
  referralBonus: number;
  platformFeePercent: number;
  minWithdrawal: number;
  minDeposit: number;
}

export interface AdminStats {
  totalUsers: number;
  totalMatches: number;
  pendingDeposits: number;
  pendingWithdrawals: number;
  totalRevenue: number;
  activeMatches: number;
  totalDeposited: number;
  totalWithdrawn: number;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  name: string;
  phone: string;
  password: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface JoinMatchInput {
  roomId: number;
}

export interface CancelMatchInput {
  matchId: string;
}

export interface DepositInput {
  amount: number;
  trxId: string;
  method: string;
  phone: string;
}

export interface WithdrawalInput {
  amount: number;
  method: string;
  phone: string;
}

export interface RoomInput {
  name: string;
  label: string;
  entryFee: number;
  winnerPrize: number;
  tagline: string;
  color: string;
}

export interface RoomUpdate {
  name?: string;
  label?: string;
  entryFee?: number;
  winnerPrize?: number;
  tagline?: string;
  color?: string;
  active?: boolean;
}

export interface BanInput {
  banned: boolean;
}

export interface BalanceEditInput {
  mainBalance: number;
  winningBalance: number;
  bonusBalance: number;
}

export interface NotificationInput {
  /** @nullable */
  userId?: number | null;
  title: string;
  message: string;
}

export interface AppSettingsInput {
  telegramLink?: string;
  referralBonus?: number;
  platformFeePercent?: number;
  minWithdrawal?: number;
  minDeposit?: number;
}