# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Royal Ludo Arena API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/admin-login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login (restricted to specific email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "403":
          description: Not authorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /deposits:
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Create deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
    get:
      operationId: getDeposits
      tags: [deposits]
      summary: Get deposits for current user
      responses:
        "200":
          description: Deposits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
...
  /admin/users/{id}/ban:
    patch:
      operationId: banUser
      tags: [admin]
      summary: Ban or unban a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: User ban status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/balance:
    patch:
      operationId: editUserBalance
      tags: [admin]
      summary: Edit user balance (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceEditInput"
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
...
  /admin/deposits/{id}/approve:
    patch:
      operationId: approveDeposit
      tags: [admin]
      summary: Approve a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/deposits/{id}/reject:
    patch:
      operationId: rejectDeposit
      tags: [admin]
      summary: Reject a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
  /admin/rooms:
    post:
      operationId: createRoom
      tags: [admin]
      summary: Create a game room (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /admin/rooms/{id}:
    patch:
      operationId: updateRoom
      tags: [admin]
      summary: Update a game room (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [admin]
      summary: Delete a game room (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        referralCode:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        banned:
          type: boolean
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required: [id, name, phone, referralCode, wins, losses, banned, isAdmin, createdAt]
    Wallet:
      type: object
      properties:
        userId:
          type: integer
        mainBalance:
          type: number
        winningBalance:
          type: number
        bonusBalance:
          type: number
        totalBalance:
          type: number
      required: [userId, mainBalance, winningBalance, bonusBalance, totalBalance]
    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        label:
          type: string
        entryFee:
          type: number
        winnerPrize:
          type: number
        playerCount:
          type: integer
        tagline:
          type: string
        color:
          type: string
        active:
          type: boolean
      required: [id, name, label, entryFee, winnerPrize, playerCount, tagline, color, active]
...
    Deposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        trxId:
          type: string
        method:
          type: string
        phone:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required: [id, userId, userName, amount, trxId, method, phone, status, createdAt]
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        referralCode:
          type: ["string", "null"]
      required: [name, phone, password]
    LoginInput:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required: [phone, password]
    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
...
    DepositInput:
      type: object
      properties:
        amount:
          type: number
        trxId:
          type: string
        method:
          type: string
        phone:
          type: string
      required: [amount, trxId, method, phone]
...
    RoomInput:
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        entryFee:
          type: number
        winnerPrize:
          type: number
        tagline:
          type: string
        color:
          type: string
      required: [name, label, entryFee, winnerPrize, tagline, color]
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        entryFee:
          type: number
        winnerPrize:
          type: number
        tagline:
          type: string
        color:
          type: string
        active:
          type: boolean
    BanInput:
      type: object
      properties:
        banned:
          type: boolean
      required: [banned]
    BalanceEditInput:
      type: object
      properties:
        mainBalance:
          type: number
        winningBalance:
          type: number
        bonusBalance:
          type: number
      required: [mainBalance, winningBalance, bonusBalance]
...