# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Royal Ludo Arena API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  label: string;
  entryFee: number;
  winnerPrize: number;
  playerCount: number;
  tagline: string;
  color: string;
  active: boolean;
}