# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Royal Ludo Arena API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  referralCode: string;
  wins: number;
  losses: number;
  banned: boolean;
  isAdmin: boolean;
  createdAt: string;
}